allow adding a tip "above" root
Allow adding an outgroup tip which diverged before the ingroup MRCA.the `ape::bind.tree` function which `bind.tip` wraps allows this, but the tree to which the tip is added ("x") must have a `root.edge`. The code submitted here does the necessary edits so that `bind.tree` does what is requested with no more arguments required from the user.